Rename eere calculations file to impacts
